requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
cfgv==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.8
Django==5.1
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-rest-passwordreset==1.4.1
djangorestframework==3.15.2
drf-yasg==1.21.7
filelock==3.15.4
identify==2.6.0
inflection==0.5.1
kombu==5.4.0
nodeenv==1.9.1
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pre-commit==3.8.0
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
ruff==0.5.6
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
